clients/publishing: refactor publishing settings file
