+++ clients/packages/polarkit/package.json
@@ -36,6 +36,7 @@
     "./components/ui/popover": "./src/components/ui/popover.tsx",
     "./components/ui/textarea": "./src/components/ui/textarea.tsx",
     "./components/ui/tabs": "./src/components/ui/tabs.tsx",
+    "./components/ui/radio-group": "./src/components/ui/radio-group.tsx",
     "./components/ui/separator": "./src/components/ui/separator.tsx",
     "./components/ui/table": "./src/components/ui/table.tsx",
     "./components/ui/accordion": "./src/components/ui/accordion.tsx",
@@ -63,6 +64,7 @@
     "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-popover": "^1.0.5",
+    "@radix-ui/react-radio-group": "^1.1.3",
     "@radix-ui/react-select": "^1.2.2",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
